various networking tweaks
