+++ backend/src/avahi.ts
@@ -55,11 +55,6 @@ export class AvahiService {
     }
   }
 
-  static cleanup(): void {
-    debug('Cleaning up Avahi service');
-    this.stopAdvertisedService();
-  }
-
   /**
    * Discovers HTTP services on the local network.
    * @returns A promise resolving to an array of discovered services.
@@ -103,18 +98,5 @@ export class AvahiService {
     }
 
     return services;
-  }
-
-  // Checks if there is any network interface 'UP'.
-  static async hasOnlineInterface(): Promise<boolean> {
-    const command = 'ip link show | grep "state UP"';
-    try {
-      const { stdout, stderr } = await execPromise(command);
-      debug(`ip link show stdout: ${stdout}`);
-      return stdout.length > 0;
-    } catch (error) {
-      debug(`Error running ip link show: ${error}`);
-      return false;
-    }
   }
 }
